Classification/PytorchTemplate/requirements.txt
cycler==0.10.0
decorator==4.4.2
future==0.18.2
imagecodecs==2020.2.18
imageio==2.8.0
joblib==0.14.1
kiwisolver==1.2.0
matplotlib==3.2.1
networkx==2.4
numpy==1.18.4
pandas==1.0.3
Pillow==7.1.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
scikit-image==0.17.2
scikit-learn==0.23.0
scipy==1.4.1
seaborn==0.10.1
six==1.14.0
threadpoolctl==2.0.0
tifffile==2020.5.11
torch==1.5.0
torchvision==0.6.0
tqdm==4.46.0
